python-webapp/requirements.txt
tornado
matplotlib
numpy
pandas
Pillow
scikit-build
torch==1.13
wandb
scikit-image
